requirements.txt
fastapi
uvicorn
redis
cryptography
pydantic
pydantic-settings
httpx
pyyaml
pytest
